requirements.txt
# BLOCKCHAIN

aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
bitarray==3.4.2
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.8.2
cytoolz==1.0.1
dj-rest-auth==7.0.1
Django==4.2.22
django-allauth==65.9.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.28.0
eth-abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eventlet==0.40.0
frozenlist==1.6.0
greenlet==3.2.2
hexbytes==1.3.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
multidict==6.4.4
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pyunormalize==16.0.0
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20250602
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
validate-docbr==1.11.0
vine==5.1.0
wcwidth==0.2.13
web3==7.12.0
websockets==15.0.1
yarl==1.20.0
